Cartridge RAM working
Rolled back MBC2 behavior so that it is now passing all mooneye-gb tests but one: MBC2/bits_ramg fails. But this was also the fact for a long time at MiSTer (they fixed it in the meantime). And during this time, there was no known MBC2 game failing due to this. So for now it is not worth the effort for me to proceed here: I consider the system now as "stable enough" and therefore *very* compatible with all MBCs including RAM MBCs.