double buffering works
After having fixed "if vga_row_next = 0 then" (which was at the last line before) already in the last commit, double buffering works now. It can be seen as now certain artefacts are not visible any more in e.g. "Grid Scroll" in the gb240p test suite. But still, the flickering on PAL @ 50 Hz remains.